rename: 이벤트 페이지 관련 파일 정리
