+++ src/app/routes/Router.tsx
@@ -4,7 +4,7 @@ import Layout from '../Layout';
 import AgreementPage from '../../pages/join/AgreementPage';
 import InfoInputPage from '../../pages/join/InfoInputPage';
 import MainPage from '../../pages/home/ui/MainPage';
-import FunnelPage from '../../pages/event-manage/ui/FunnelPage';
+import FunnelPage from '../../pages/event/ui/create-event/FunnelPage';
 import AllEventsPage from '../../pages/event/ui/AllEventsPage';
 import MyTicketPage from '../../pages/menu/ui/MyTicketPage';
 import SearchPage from '../../pages/search/ui/SearchPage';
