+++ src/features/event/create-event/ui/EventFunnel.tsx
@@ -1,14 +1,14 @@
-import HostSelectionPage from '../../../pages/event/ui/host/HostSelectionPage';
-import HostCreationPage from '../../../pages/event/ui/host/HostCreationPage';
-import EventTitlePage from '../../../pages/event/ui/create-event/EventTitlePage';
-import EventPeriodPage from '../../../pages/event/ui/create-event/EventPeriodPage';
-import EventInfoPage from '../../../pages/event/ui/create-event/EventInfoPage';
-import EventTypePage from '../../../pages/event/ui/create-event/EventTypePage';
-import EventTagPage from '../../../pages/event/ui/create-event/EventTagPage';
-import EventOrganizerInfoPage from '../../../pages/event/ui/create-event/EventOrganizerInfoPage';
-import EventRegisterLayout from '../../../shared/ui/backgrounds/EventRegisterLayout';
+import HostSelectionPage from '../../../../pages/event/ui/host/HostSelectionPage';
+import HostCreationPage from '../../../../pages/event/ui/host/HostCreationPage';
+import EventTitlePage from '../../../../pages/event/ui/create-event/EventTitlePage';
+import EventPeriodPage from '../../../../pages/event/ui/create-event/EventPeriodPage';
+import EventInfoPage from '../../../../pages/event/ui/create-event/EventInfoPage';
+import EventTypePage from '../../../../pages/event/ui/create-event/EventTypePage';
+import EventTagPage from '../../../../pages/event/ui/create-event/EventTagPage';
+import EventOrganizerInfoPage from '../../../../pages/event/ui/create-event/EventOrganizerInfoPage';
+import EventRegisterLayout from '../../../../shared/ui/backgrounds/EventRegisterLayout';
 import { useNavigate } from 'react-router-dom';
-import { EventFunnelInterface, StepNames } from '../../../shared/types/funnelType';
+import { EventFunnelInterface, StepNames } from '../../../../shared/types/funnelType';
 import { useFunnelState } from '../model/FunnelContext';
 import { useEventCreation } from '../hooks/useEventHook';
 import { useHostCreation } from '../hooks/useHostHook';
