+++ src/pages/event/ui/create-event/EventInfoPage.tsx
@@ -1,7 +1,7 @@
-import FileUpload from '../../../features/event/ui/FileUpload';
-import TextEditor from '../../../features/event/ui/TextEditor';
-import LinkInput from '../../../features/event/ui/LinkInput';
-import { useFunnelState } from '../../../features/event/model/FunnelContext';
+import FileUpload from '../../../../features/event/ui/FileUpload';
+import TextEditor from '../../../../features/event/ui/TextEditor';
+import LinkInput from '../../../../features/event/ui/LinkInput';
+import { useFunnelState } from '../../../../features/event/model/FunnelContext';
 
 const EventInfoPage = () => {
   const { setEventState } = useFunnelState();
